Port to Python 3

+++ controls.py
@@ -313,7 +313,7 @@ class Controls(GObject.GObject):
 
     def _reseek(self):
         self._scale_reseek_timeout_id = -1
-        location = long(self.activity.control.hscale.get_value() *
+        location = int(self.activity.control.hscale.get_value() *
                         self.p_duration / 100)  # in ns
         self.activity.player.seek(location)
         # Allow for a preroll
